tox/travis: fix environment setup and un-vendored support
Use the virtualenv's pip for installing dependencies, not the
code that is about to be tested.

Add 2 dedicated un-vendored environments (py27-novendor and
py36-novendor), making it easier to use locally for testing.

Un-vendoring is done by using a custom `install_command` that
automatically un-vendors pip each time it's installed.

commit 922f9a3f6dc4ac642492204808b4d5cc995aca66
Author: Benoit Pierre <benoit.pierre@gmail.com>
Date:   Wed Aug 9 02:47:39 2017 +0200

    tox/travis: fix and rework handling of un-vendored jobs