tests: rework Windows workaround for maximum path length limitation
The new pytest release (3.9.1) switch to using `pathlib.Path.resolve`
on Python 3 means mapping `TMP`/`TEMP` to a temporary drive does not
work anymore (the path is expanded back to the full mapped-to location),
additionally, the length of pytest base temporary directory path has
increased (thanks to another component including the user's name), which
means that some of our tests ran afoul of the maximum path length
limitation on Python 2 too.

The solution: use a custom `tmpdir_factory` fixture that shorten the
paths back to a temporary mapped drive when using the new testsuite
option: `--tmpdir-subst-drive`.